+++ components/keyemulate/impl/src/main/java/com/flipperdevices/keyemulate/viewmodel/EmulateViewModel.kt
@@ -109,7 +109,9 @@ abstract class EmulateViewModel(
             return
         }
         if (!emulateStarted) {
+            info { "Failed start emulation" }
             emulateHelper.stopEmulateForce(requestApi)
+            emulateButtonStateFlow.emit(EmulateButtonState.Inactive())
         }
     }