+++ components/keyscreen/impl/src/test/kotlin/com/flipperdevices/keyscreen/impl/viewmodel/KeyScreenViewModelTest.kt
@@ -115,7 +115,8 @@ class KeyScreenViewModelTest {
                 deleteState = mockk(),
                 flipperKey = flipperKey,
                 emulateConfig = mockk(),
-                isSupportEditing = true
+                isSupportEditing = true,
+                emulatingInProgress = false
             )
         )